Freeze ECS pipelines at tagged last commit on ECS project.
Folks want to keep the ECS stuff around for a bit in the test account,
but we don't want Concourse trying to apply subsequent Terraform changes
to it.

The apps should continue to be updated/deployed, but the TF will stay at
the tag ecs-latest-before-pivot-to-k8s.

This will allow us to declutter the repo by removing the old ECS stuff
as we continue to work on the Kubernetes project.

Rollout:

    fly sp -t govuk-test -p deploy-apps -c concourse/pipelines/deploy.yml -l concourse/parameters/test/deploy.yml

and similarly for the `monitoring` pipeline.