Adding Nginx reverse proxy to Rails apps as sidecar.
We're not 100% sure what behaviour these will need in the future. Possibly they'll be involved in serving static assets  such ad CSS, JS etc.

It will certainly be needed in some applications such asset manager, which uses ‘x-accel-redirect’.
We are not mandating the use of Nginx sidecar for every Rails app, though we believe it will be the case.
If it turns out that the Nginx sidecar have no behaviour, we may consider removing them at a later date.

Some of the reasons to why we are adding Nginx sidecar:

  We want apps to be able to work as independently as possible.

  Helm chart should be a self-contained encapsulated artefact.

  Nginx sidecar will just be doing reverse proxy, static serving and possibly some x-accel-redirect stuff to s3 for the few apps that need this.

  Unicorn can't deal with slow clients efficiently.

Nginx sidecar Config:

  The Nginx config itself will need to be updated to support functionality we require in the future,

  Some reference for NGINX requirements can be found here:
  https://docs.google.com/document/d/1hOgkeimIn20O5zrD0wTflap1_DTVatb7NOnb223NhSE/
  https://trello.com/c/Lkb8N313/381-epic-nginx-and-varnish-changes-needed

Values naming changes:

  Rails application image to: appImage.
  Rails application container port to: appPort.
  Nginx sidecar image to - nginxImage.
  Nginx sidecar container portto: nginxPort.

Apps that now contain Nginx sidecar:

  Content-store.
  Frontend.
  Static.

All of the above three apps now contain:

  New values in value.yaml.
  Nginx container and volume in template/deployment.yaml.
  Configmap which contains nginx.conf.